% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyorthologs.R
\name{shinyorthologs}
\alias{shinyorthologs}
\title{Launch the shinyorthologs app}
\usage{
shinyorthologs(db_user = NULL, db_host = NULL, db_port = NULL,
  db_pass = NULL, db_name = NULL)
}
\arguments{
\item{db_user}{Database user}

\item{db_host}{Database host}

\item{db_port}{Database port}

\item{db_pass}{Database password}

\item{db_name}{Database name}
}
\description{
Executing this function will launch the shinyorthologs application in
the user's default web browser.
}
\examples{
\dontrun{
shinyorthologs()
}
}
\author{
Colin Diesh \email{dieshcol@msu.edu}
}
