% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyorthologs.R
\name{shinyorthologs}
\alias{shinyorthologs}
\title{Launch the shinyorthologs app}
\usage{
shinyorthologs(user = NULL, host = NULL, port = NULL, password = NULL,
  dbname = NULL, basedir = NULL, dev = F)
}
\arguments{
\item{user}{Database user}

\item{host}{Database host}

\item{port}{Database port}

\item{password}{Database password}

\item{dbname}{Database name}

\item{basedir}{Root directory for fasta/expression files}

\item{dev}{Boolean if using dev environment, loads from local directories}
}
\description{
Executing this function will launch the shinyorthologs application in
the user's default web browser.
}
\examples{
\dontrun{
shinyorthologs(basedir='/data/dir', dbname = 'shinyorthologs')
}
}
