% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyorthologs.R
\name{shinyorthologs}
\alias{shinyorthologs}
\title{Launch the shinyorthologs app}
\usage{
shinyorthologs()
}
\description{
Executing this function will launch the shinyorthologs application in
the user's default web browser.
}
\examples{
\dontrun{
shinyorthologs()
}
}
\author{
Colin Diesh \email{dieshcol@msu.edu}
}
